+++ README.en.md
@@ -23,6 +23,13 @@ Project Memory v2 kernel, respectively. The root package exposes those surfaces 
 facades and no session data needs migration;
 see the [modular architecture](docs/en/architecture.md).
 
+In the **0.63 Harness FLOW Store** phase, `packages/vault/src/locale.mjs` becomes the canonical
+source for Vault locale and taxonomy, while `packages/harness/src/flow-store.mjs` becomes the
+canonical durable FLOW store. The historical `hooks/locale.mjs` and
+`hooks/vault-runtime-store.mjs` facades preserve export identity. Harness depends only on Vault's
+public index — never the reverse — with no migration of paths, schemas, or locks; the tarball
+remains a single artifact published by the root `wendkeep` package.
+
 ```bash
 npm i -D wendkeep && npx wendkeep init      # captures from the next session on
 npx wendkeep import                          # backfill past Claude + Codex sessions
