+++ packages/harness/src/index.mjs
@@ -1,2 +1,3 @@
+export * from './flow-store.mjs';
 export * from './operating-profile.mjs';
 export * from './sensors-core.mjs';
