+++ packages/vault/src/index.mjs
@@ -1,5 +1,6 @@
 export * from './project-vault.mjs';
 export * from './vault-path-safety.mjs';
+export * from './locale.mjs';
 export * from './memory-schema.mjs';
 export * from './memory-mode.mjs';
 export * from './memory-handoff.mjs';
