+++ tests/tarball-smoke.test.mjs
@@ -73,11 +73,12 @@ test('every published hook passes node --check (no broken syntax shipped)', () =
   }
 });
 
-test('[req:MOD-4] [req:MOD-8] [req:MOD-10] published tarball contains the modular Vault and Harness surfaces', () => {
+test('[req:MOD-4] [req:MOD-8] [req:MOD-10] [req:MOD-11] [req:MOD-12] [req:MOD-13] published tarball contains the modular Vault and Harness surfaces', () => {
   const published = publishedFiles();
   for (const path of [
     'packages/cli/package.json',
     'packages/harness/package.json',
+    'packages/harness/src/flow-store.mjs',
     'packages/harness/src/index.mjs',
     'packages/harness/src/operating-profile.mjs',
     'packages/harness/src/sensors-core.mjs',
@@ -86,6 +87,7 @@ test('[req:MOD-4] [req:MOD-8] [req:MOD-10] published tarball contains the modula
     'packages/pi/package.json',
     'packages/vault/package.json',
     'packages/vault/src/index.mjs',
+    'packages/vault/src/locale.mjs',
     'packages/vault/src/memory-handoff.mjs',
     'packages/vault/src/memory-mode.mjs',
     'packages/vault/src/memory-schema.mjs',
@@ -99,7 +101,7 @@ test('[req:MOD-4] [req:MOD-8] [req:MOD-10] published tarball contains the modula
   }
 });
 
-test('[req:MOD-4] [req:MOD-6] [req:MOD-9] [req:MOD-10] installed tarball exposes Harness identities with OFF and Vault intact', () => {
+test('[req:MOD-4] [req:MOD-6] [req:MOD-9] [req:MOD-10] [req:MOD-11] [req:MOD-12] [req:MOD-13] installed tarball exposes strict canonical, public, and legacy identities', () => {
   const temp = mkdtempSync(join(tmpdir(), 'wendkeep-installed-tarball-'));
   try {
     const packed = spawnSync(`npm pack --json --pack-destination "${temp}"`, {
@@ -126,8 +128,12 @@ test('[req:MOD-4] [req:MOD-6] [req:MOD-9] [req:MOD-10] installed tarball exposes
     const imported = spawnSync(process.execPath, ['--input-type=module', '--eval', [
       "const vault = await import('wendkeep/vault');",
       "const harness = await import('wendkeep/harness');",
+      "const canonicalLocale = await import('wendkeep/packages/vault/src/locale.mjs');",
+      "const canonicalFlowStore = await import('wendkeep/packages/harness/src/flow-store.mjs');",
+      "const legacyLocale = await import('wendkeep/hooks/locale.mjs');",
       "const legacy = await import('wendkeep/hooks/vault-path-safety.mjs');",
       "const legacyStore = await import('wendkeep/hooks/memory-store.mjs');",
+      "const legacyFlowStore = await import('wendkeep/hooks/vault-runtime-store.mjs');",
       "const legacyProfile = await import('wendkeep/src/operating-profile.mjs');",
       "const legacySensors = await import('wendkeep/hooks/sensors-core.mjs');",
       "if (typeof vault.resolveProjectVault !== 'function') process.exit(11);",
@@ -144,6 +150,63 @@ test('[req:MOD-4] [req:MOD-6] [req:MOD-9] [req:MOD-10] installed tarball exposes
       "const off = harness.operatingProfilePolicy('OFF');",
       "if (off.keepCore !== true || off.harness !== false || off.contract !== 'native') process.exit(21);",
       "if (vault.canonicalMemoryJson({ z: 1, a: 2 }) !== '{\"a\":2,\"z\":1}') process.exit(22);",
+      "for (const [name, value] of Object.entries(canonicalLocale)) {",
+      "  if (vault[name] !== value || legacyLocale[name] !== value) process.exit(23);",
+      "}",
+      "for (const [name, value] of Object.entries(canonicalFlowStore)) {",
+      "  if (harness[name] !== value || legacyFlowStore[name] !== value) process.exit(24);",
+      "}",
+      "const { join } = await import('node:path');",
+      "const { mkdirSync } = await import('node:fs');",
+      "const runtimeVault = join(process.cwd(), 'consumer-vault');",
+      "mkdirSync(runtimeVault, { recursive: true });",
+      "const contract = (flowId, sessionId) => ({",
+      "  schema_version: 1, flow_id: flowId, session_id: sessionId,",
+      "  session_file: '02-Sessões/installed.md', slug: `slug-${flowId}`, profile: 'FLOW',",
+      "  started_at: '2026-07-28T12:00:00.000Z', reason: 'installed tarball proof',",
+      "  spec_impact: 'none', project_rel: '.', protected_roots: [],",
+      "  allowed_paths: ['src/a.mjs'], sensor_ids: ['tests'], sensor_definition_hash: 'hash',",
+      "  baseline: { schema_version: 1, root: 'C:/consumer', head: 'abc', fingerprints: {},",
+      "    git_metadata_fingerprint: 'a'.repeat(64), hidden_index_paths: [],",
+      "    unsafe_git_metadata_paths: [], unsafe_worktree_paths: [] },",
+      "});",
+      "const finishedContract = contract('flow-finished', 'session-finished');",
+      "canonicalFlowStore.createFlowContract(runtimeVault, finishedContract);",
+      "canonicalFlowStore.appendFlowAttempt(runtimeVault, 'session-finished', 'flow-finished', {",
+      "  schema_version: 1, attempt_id: 'attempt-1', status: 'red',",
+      "  recorded_at: '2026-07-28T12:01:00.000Z', failures: ['expected red'],",
+      "  changed_paths: ['src/a.mjs'], evidence: [{ id: 'tests', status: 'red',",
+      "    ts: '2026-07-28T12:01:00.000Z', severity: 'critical' }],",
+      "});",
+      "canonicalFlowStore.writeFlowReceipt(runtimeVault, 'session-finished', 'flow-finished', {",
+      "  schema_version: 1, flow_id: 'flow-finished', status: 'finished',",
+      "  finished_at: '2026-07-28T12:02:00.000Z', reason: finishedContract.reason,",
+      "  allowed_paths: finishedContract.allowed_paths, sensor_ids: finishedContract.sensor_ids,",
+      "  changed_paths: ['src/a.mjs'], evidence: [{ id: 'tests', status: 'green',",
+      "    ts: '2026-07-28T12:02:00.000Z', severity: 'critical' }],",
+      "  baseline_head: 'abc', final_head: 'abc',",
+      "});",
+      "const finished = canonicalFlowStore.readFlow(runtimeVault, { sessionId: 'session-finished', flowId: 'flow-finished' });",
+      "if (finished.state !== 'finished' || finished.attempts.length !== 1) process.exit(25);",
+      "const promotedContract = contract('flow-promoted', 'session-promoted');",
+      "canonicalFlowStore.createFlowContract(runtimeVault, promotedContract);",
+      "canonicalFlowStore.reserveFlowPromotion(runtimeVault, 'session-promoted', 'flow-promoted', {",
+      "  schema_version: 1, flow_id: 'flow-promoted', status: 'promoting',",
+      "  reserved_at: '2026-07-28T12:03:00.000Z', change_slug: 'installed-change',",
+      "  change_rel: '08-Mudanças/installed-change', origin: { schema_version: 1,",
+      "    flow_id: 'flow-promoted', promoted_at: '2026-07-28T12:03:00.000Z',",
+      "    contract: promotedContract, attempts: [], observed_git: { baseline_head: 'abc',",
+      "      current_head: 'abc', head_changed: false, changed_paths: ['src/a.mjs'] } },",
+      "});",
+      "canonicalFlowStore.writeFlowPromotion(runtimeVault, 'session-promoted', 'flow-promoted', {",
+      "  schema_version: 1, flow_id: 'flow-promoted', status: 'promoted',",
+      "  promoted_at: '2026-07-28T12:03:00.000Z', change_slug: 'installed-change',",
+      "  change_rel: '08-Mudanças/installed-change',",
+      "  origin_file: '08-Mudanças/installed-change/flow-origin.json',",
+      "  changed_paths: ['src/a.mjs'], baseline_head: 'abc', current_head: 'abc',",
+      "});",
+      "const promoted = canonicalFlowStore.readFlow(runtimeVault, { sessionId: 'session-promoted', flowId: 'flow-promoted' });",
+      "if (promoted.state !== 'promoted' || canonicalFlowStore.listFlows(runtimeVault).length !== 2) process.exit(26);",
     ].join('\n')], { cwd: consumer, encoding: 'utf8' });
     assert.equal(imported.status, 0, `installed imports failed:\n${imported.stderr}`);
